requirements.txt
numpy
pandas
scikit-learn
streamlit
google-generativeai
pickle